modify

+++ utils_.py
@@ -41,35 +41,54 @@ def mkdir_join(dir):
 def prefetch_srav2(sralist,outdir):
     #ss = " ".join(sralist)
     #progress_bar("prefetch")
-    outdir
-    try:
-        cmd = "singularity exec --containall --bind /work/linsslab01/:/home/linsslab01/ /work/linsslab01/sra-tools_latest.sif prefetch "+sralist+" --output-directory "+outdir.replace("work","home")
-        print (cmd,"\n")
-        run_cmd3(cmd)
-        time.sleep(1)
-    except Exception as e:
-        print ("prefetch has problem:\n")
-        error_class = e.__class__.__name__  # 取得錯誤類型
-        detail = e.args[0]  # 取得詳細內容
-        cl, exc, tb = sys.exc_info()  # 取得Call Stack
-        lastCallStack = traceback.extract_tb(tb)[-1]  # 取得Call Stack的最後一筆資料
-        fileName = lastCallStack[0]  # 取得發生的檔案名稱
-        lineNum = lastCallStack[1]  # 取得發生的行號
-        funcName = lastCallStack[2]  # 取得發生的函數名稱
-        errMsg = "File \"{}\", line {}, in {}: [{}] {}".format(fileName, lineNum, funcName, error_class, detail)
-        print(errMsg)
-        with open("./SRA_run_error.txt", "a+") as f:
-            f.write("{} :\n{}\n".format(sralist, errMsg))
-        #處理檔案已存在問題
-        current_path = os.path.join(os.path.abspath(os.getcwd()), sralist)
-        print(current_path, "\n")
-        print("shutil.rmtree({})\n".format(current_path))
-        shutil.rmtree(current_path)
+    sra_dir=os.path.join(outdir,sralist)
+    sra_file = os.path.join(outdir,"{}.sra".format(sralist))
+    if os.path.isfile(sra_file):
+        print("{} have .sra file\n".format(sra_dir))
+    else:
+        try:
+            cmd = "singularity exec --containall --bind /work/linsslab01/:/home/linsslab01/ /work/linsslab01/sra-tools_latest.sif prefetch "+sralist+" --output-directory "+outdir.replace("work","home")
+            print (cmd,"\n")
+            run_cmd3(cmd)
+            time.sleep(1)
+        except Exception as e:
 
-        print("######run again\n")
-        run_cmd2(cmd)
-        time.sleep(1)
-        pass
+            print ("prefetch has problem:\n")
+            error_class = e.__class__.__name__  # 取得錯誤類型
+            detail = e.args[0]  # 取得詳細內容
+            cl, exc, tb = sys.exc_info()  # 取得Call Stack
+            lastCallStack = traceback.extract_tb(tb)[-1]  # 取得Call Stack的最後一筆資料
+            fileName = lastCallStack[0]  # 取得發生的檔案名稱
+            lineNum = lastCallStack[1]  # 取得發生的行號
+            funcName = lastCallStack[2]  # 取得發生的函數名稱
+            errMsg = "File \"{}\", line {}, in {}: [{}] {}".format(fileName, lineNum, funcName, error_class, detail)
+            print(errMsg)
+            with open("./prefetch_err.txt", "a+") as f:
+                f.write("{} :\n{}\n".format(sralist, errMsg))
+            #處理檔案已存在問題
+            current_path = os.path.join(os.path.abspath(os.getcwd()), sralist)
+            print(current_path, "\n")
+            print("shutil.rmtree({})\n".format(current_path))
+            shutil.rmtree(current_path)
+
+            print("######run again\n")
+            try:
+                run_cmd2(cmd)
+            except Exception as e:
+                print("prefetch has problem:\n")
+                error_class = e.__class__.__name__  # 取得錯誤類型
+                detail = e.args[0]  # 取得詳細內容
+                cl, exc, tb = sys.exc_info()  # 取得Call Stack
+                lastCallStack = traceback.extract_tb(tb)[-1]  # 取得Call Stack的最後一筆資料
+                fileName = lastCallStack[0]  # 取得發生的檔案名稱
+                lineNum = lastCallStack[1]  # 取得發生的行號
+                funcName = lastCallStack[2]  # 取得發生的函數名稱
+                errMsg = "File \"{}\", line {}, in {}: [{}] {}".format(fileName, lineNum, funcName, error_class, detail)
+                print(errMsg)
+                with open("./prefetch_err.txt", "a+") as f:
+                    f.write("{} :\n{}\n".format(sralist, errMsg))
+            time.sleep(1)
+            pass
     print("now download", sralist, "runs.")
